chore: Add check for prehandle-handle consistency (#21285)

+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/container/ContainerTransactionGenerator.java
@@ -106,7 +106,7 @@ public class ContainerTransactionGenerator implements TransactionGenerator {
 
         for (final Node node : nodes) {
             if (node.platformStatus() == PlatformStatus.ACTIVE) {
-                final byte[] transaction = TransactionFactory.createEmptyTransaction(random.nextInt())
+                final byte[] transaction = TransactionFactory.createEmptyTransaction(random.nextLong())
                         .toByteArray();
                 try {
                     node.submitTransaction(transaction);
