+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/turtle/TurtleTransactionGenerator.java
@@ -85,7 +85,7 @@ public class TurtleTransactionGenerator implements TransactionGenerator {
             for (long i = previousCount; i < currentCount; i++) {
                 for (final Node node : activeNodes) {
                     // Generate a random transaction and submit it to the node.
-                    final byte[] transaction = TransactionFactory.createEmptyTransaction(randotron.nextInt())
+                    final byte[] transaction = TransactionFactory.createEmptyTransaction(randotron.nextLong())
                             .toByteArray();
                     node.submitTransaction(transaction);
                 }
